move definitions of new functions/interfaces to the original file
